requirements.txt
yfinance==0.2.18
pandas==1.5.3
numpy==1.24.3
newsapi-python==0.2.7
scikit-learn==1.2.2
statsmodels==0.13.5
matplotlib==3.7.1
seaborn==0.12.2
streamlit==1.22.0
lightgbm==4.4.0
textblob==0.18.0
plotly==5.22.0
pandas_market_calendars==4.4.1